refactor: 로그인 팝업 로직을 전용 훅으로 분리
